[add] added backends/util.h
