+++ src/backends/torch.c
@@ -1,5 +1,5 @@
 #include "backends/torch.h"
-#include "backends.h"
+#include "backends/util.h"
 #include "tensor.h"
 #include "util/arr_rm_alloc.h"
 #include "libtorch_c/torch_c.h"
